refactor: move epoch_number_with_fraction_cmp to utils.h
